src/test/java/com/sedykh/trie/DesignAddAndSearchWordsDataStructureTest.java
package com.sedykh.trie;

import org.junit.jupiter.api.Test;
import static org.junit.jupiter.api.Assertions.assertFalse;
import static org.junit.jupiter.api.Assertions.assertTrue;

class DesignAddAndSearchWordsDataStructureTest {


    @Test
    public void test() {
        DesignAddAndSearchWordsDataStructure.WordDictionary wordDictionary = new DesignAddAndSearchWordsDataStructure.WordDictionary();
        wordDictionary.addWord("bad");
        wordDictionary.addWord("dad");
        wordDictionary.addWord("mad");
        assertFalse(wordDictionary.search("pad")); // return False
        assertTrue(wordDictionary.search("bad")); // return True
        assertTrue(wordDictionary.search(".ad")); // return True
        assertTrue(wordDictionary.search("b..")); // return True
    }
}